Store money as Decimal, add Print support to reports
All monetary/quantity model columns (Product prices, order/bill/invoice
lines, payments, budget lines) move from Float to Numeric(12,2)/Decimal
so totals and balances never accumulate float rounding error. Every
router that parses form input into these fields now builds Decimal
values instead of float, and two comparisons that combined a raw float
literal (0.01) with a now-Decimal amount (which raises TypeError) are
fixed to compare against Decimal("0.01") instead, in the vendor-bill
and customer-invoice overpayment/paid-status checks.

Also adds a Print button and a `@media print` stylesheet to the
Balance Sheet, Profit & Loss, and Budget Report pages: printing hides
the nav/filter controls and drill-down links and shows a clean report
header with the report name, filter period, and generation timestamp
instead.

Verified with the full 107-test suite (unchanged, all green) plus a
manual TestClient smoke check of all three report pages.

+++ app/routers/products.py
@@ -1,4 +1,5 @@
 import os
+from decimal import Decimal
 
 from fastapi import APIRouter, Depends, File, Form, Request, UploadFile
 from fastapi.responses import RedirectResponse
@@ -58,8 +59,8 @@ def create_product(
     request: Request,
     name: str = Form(...),
     type: str = Form(...),
-    sales_price: float = Form(...),
-    cost_price: float = Form(...),
+    sales_price: Decimal = Form(...),
+    cost_price: Decimal = Form(...),
     category: str = Form(""),
     image: UploadFile | None = File(None),
     user: User = Depends(require_role("admin", "accountant")),
@@ -132,8 +133,8 @@ def update_product(
     request: Request,
     name: str = Form(...),
     type: str = Form(...),
-    sales_price: float = Form(...),
-    cost_price: float = Form(...),
+    sales_price: Decimal = Form(...),
+    cost_price: Decimal = Form(...),
     category: str = Form(""),
     image: UploadFile | None = File(None),
     user: User = Depends(require_role("admin")),
